requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
mysql-connector==2.2.9
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1

